fix: Button 활성화 로직 변경
